easy looking
removed bottom navi class
add header to each component
add newsletter to bottombar

+++ alpin_frontend/lib/components/suggestion/suggestion_view.dart
@@ -1,6 +1,8 @@
 import 'package:alpin_frontend/components/base_widget.dart';
 import 'package:alpin_frontend/components/suggestion/suggestion_model.dart';
+import 'package:alpin_frontend/model/trip_suggestion.dart';
 import 'package:alpin_frontend/services/language-provider/translation-service.dart';
+import 'package:alpin_frontend/widgets/customSliverAppBar.dart';
 import 'package:alpin_frontend/widgets/suggestion_widget.dart';
 import 'package:flutter/material.dart';
 import 'package:stacked/stacked.dart';
@@ -9,8 +11,7 @@ import 'package:stacked/stacked.dart';
 ///
 /// visibility act on loading tourSuggestions from server
 /// append [isLoading] variable
-class SuggestionView extends StatelessWidget implements BaseWidget{
-
+class SuggestionView extends StatelessWidget implements BaseWidget {
   @override
   String getAppBarName(context) => translation(context).tourSuggestion;
 
@@ -21,33 +22,35 @@ class SuggestionView extends StatelessWidget implements BaseWidget{
       ViewModelBuilder<SuggestionModel>.reactive(
         viewModelBuilder: () => SuggestionModel(context),
         builder: (context, model, child) {
-          return Visibility(
-            visible: model.isLoading,
-            child: ListView.builder(
-              itemBuilder: (context, index) => SuggestionWidget(
-                suggestion: model.list[index],
-                isAdmin: model.isAdmin,
-                deleteById: model.deleteSuggestionById,
-                editSuggestion: model.editSuggestion,
-              ),
-              itemCount: model.list.length,
-            ),
-            replacement: const Center(child: CircularProgressIndicator()),
+          return CustomScrollView(
+            slivers: [
+              CustomSliverAppBar(title: getAppBarName(context)),
+              SliverFillRemaining(
+                child: FutureBuilder(
+                  future: model.loadingSuggestions(),
+                  builder: (context, snapshot) {
+                    if (snapshot.hasData) {
+                      var list = snapshot.data as List<TripSuggestion>;
+                      if (list == null || list.isEmpty)
+                        return Center(
+                            child: Text(translation(context).foundNoContent));
+                      return ListView(
+                        children: [
+                          for (var item in list)
+                            SuggestionWidget(
+                                suggestion: item,
+                                isAdmin: model.isAdmin,
+                                deleteById: model.deleteSuggestionById,
+                                editSuggestion: model.editSuggestion)
+                        ],
+                      );
+                    }
+                    return const Center(child: CircularProgressIndicator());
+                  },
+                ),
+              )
+            ],
           );
-          /*return Scaffold(
-            drawer: DrawerWidget(
-                content: Config.drawerConfig(context),
-                actualRoute: RouterGenerator.suggestionView),
-            appBar: AppBar(
-              title: Text(translation(context).tourSuggestion),
-              actions: [
-                IconButton(
-                    onPressed: () => model.showDialogSuggestion(),
-                    icon: const Icon(Icons.add))
-              ],
-            ),
-            body: ,
-          );*/
         },
       );
 }
